pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newspace-hybrid-digital-twin"
version = "0.3.0"
description = "Integrated hybrid multi-sensor digital twin framework with fusion, twin state, API, and validation runtime modules."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "python-dotenv>=1.0",
  "loguru>=0.7",
  "typer>=0.12",
  "rich>=13.7",
  "pandas>=2.2",
  "numpy>=1.26",
  "geopandas>=0.14",
  "shapely>=2.0",
  "pyproj>=3.6",
  "rasterio>=1.3",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "torch>=2.2",
  "torchvision>=0.17",
  "scikit-learn>=1.4",
  "matplotlib>=3.8",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "httpx>=0.27",
]



[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "black>=24.4",
  "mypy>=1.10",
  "types-PyYAML",
  "pre-commit>=3.7",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "scripts", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "W"]
ignore = ["E203"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
